Add files via upload
In this project I have implemented Spring JDBC, JPA, and Spring Data JPA. Spring JDBC helps us to write less Java code but more SQL queries, JPA helps to map entities to tables and avoid SQL queries and Spring Data JPA is built on top of JPA, so it takes care of everything(Eg: Entity Manager) and we need use the methods in the Spring Data JPA along with defining our own methods in the interface extending JpaRepository. Hibernate is a framework that uses JPA API making it easier to switch to another framework in future.